file structure: separate app config from app listen
